requirements.txt
python-telegram-bot==20.7 --trusted-host pypi.org --trusted-host pypi.python.org --trusted-host files.pythonhosted.org
redis==5.0.1 --trusted-host pypi.org --trusted-host pypi.python.org --trusted-host files.pythonhosted.org  
mysql-connector-python==8.2.0 --trusted-host pypi.org --trusted-host pypi.python.org --trusted-host files.pythonhosted.org


